moved navbarController
