+++ Quintor/src/main/java/module-info.java
@@ -5,6 +5,4 @@ module com.example.quintor {
 
     opens com.example.quintor to javafx.fxml;
     exports com.example.quintor;
-    exports com.example.quintor.controller;
-    opens com.example.quintor.controller to javafx.fxml;
 }
